Almost done with error correction for 1.7.10

+++ src/main/scala/resonantinduction/electrical/battery/TileEnergyDistribution.java
@@ -10,7 +10,7 @@ public class TileEnergyDistribution extends TileElectric
 {
     public boolean markClientUpdate = false;
     public boolean markDistributionUpdate = false;
-    public long renderEnergyAmount = 0;
+    public double renderEnergyAmount = 0;
     private EnergyDistributionNetwork network;
 
     public TileEnergyDistribution()
